// frontend/src/gen/esteemed/v1/analytics_pb.ts
export enum DateRange {
  /**
   * @generated from enum value: DATE_RANGE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: DATE_RANGE_TODAY = 1;
   */
  TODAY = 1,

  /**
   * @generated from enum value: DATE_RANGE_LAST_7_DAYS = 2;
   */
  LAST_7_DAYS = 2,

  /**
   * @generated from enum value: DATE_RANGE_LAST_30_DAYS = 3;
   */
  LAST_30_DAYS = 3,

  /**
   * @generated from enum value: DATE_RANGE_LAST_90_DAYS = 4;
   */
  LAST_90_DAYS = 4,

  /**
   * @generated from enum value: DATE_RANGE_ALL_TIME = 5;
   */
  ALL_TIME = 5,
}